fix(Program): avoid unnecessary GitHub requests when passing over a link
Previously, the request to GitHub was initiated immediately and showing
the popup was delayed (in case the cursor was just passing over the
element). This resulted in always showing errors for invalid links (even
if the cursor was passing by).

This commit avoids "eager" errors by delaying the GitHub request and
then showing the popup immediately (as soon as the request completes).