feat(Program): make token acquisition lazy (on an as-needed basis)
Previously, the script would prompt the user for the necessary tokens at
start-up (unless they were already stored). This was creating
unnecessary noise and might end up not being useful at all (e.g. the
tokens may never be used).

This commit changes the token acquisition logic in two ways:
1. Prompting for a token is postponed, until it is actually needed.
2. The prompt does not appear automatically, but the user is offered the
   option to bring it up (with a short explanation of why the token is
   required to completed the requested operation).

This avoids unnecessary interruption for users and gives them more
context on when/what a token is needed for, allowing them to make an
informed decision.

For example, a user may never be prompted for a GitHub token, if they
never reach the rate-limit for anonymous requests. Or never be prompted
about a Jira token, if they never try to see info for a Jira issue.