fix(Program): better handle too long error messages
Previously, error messages would be included in the error snackbar
as-is. This ended up being unuseable for too large error messages (such
as Jira's returning an HTML page for unauthorized requests).

This commit fixes it by only including the first few characters of the
error message in the snackbar. The full error is still available in the
console (which is also mentioned in the snackbar).