fix(Program): check for updates even if installation failed
In some cases, it is possible that a version breaks due to some breaking
changes in Slack (e.g. in the DOM layout, element IDs/classes, etc.). To
be able to recover from such a state (without the user having to
manually look for updates) checking for updates now happens even if
installation failed.