fix(UiUtils): do not hide snackbar while it is being hovered
Previously, the snackbar was hidden after the specified delay. This did
not offer a good UX when the user intended to interact with the element
and had their mouse cursor over it.

This commit changes the hide logic so that it waits for the mouse cursor
to leave the snackbar before hiding it.