fix(Program): handle GitHub commits/PRs with large files
NgSlackLinkifier uses the `patch` property of the file objects returned
as part of commit/PR info by the GitHub API to show diffs. It turns out
that for large files (>1000 changes lines?), the GitHub API does not
include the `patch` property.

Previously, NgSlackLinkifier would fail and not display any info/preview
for a commit/PR if any of the files did not include the `patch`
property.

This commit fixes it by displaying a placeholder as diff for files that
are too large and therefore have no corresponding `patch` data sent for
them.