fix(Linkifier): avoid having encoded entities in RegExp source decoded when used as bookmarklet
When using the code as a bookmarklet, any encoded entities (e.g. `%3A`)
might be decoded by the browser in the bookmark target. This breaks the
code, therefore such entities need to remain encoded.

The previous method of using template strings and a constant for `%`
would be undone by the current version of the minifier, which would get
rid of the constant and inline `%` into the respective strings.

This commit fixes it by using a placeholder for `%` (i.e. `{{P}}`) and
doing the replacement with `%` at runtime. It also refactors `Linkifier`
to define the RegExps in one place (outside of loops) to improve
performance and make maintenance easier.