fix(Linkifier): detect Markdown links in bold or italic text
The previous logic for detecting Markdown links assumed that the various
parts of the link appear next to each other (as sibling text nodes and
anchor elements). While this is generally true, in some cases the text
nodes may be wrapped in `<b>` and/or `<i>` elements.

This commit updates the logic to be able to "unwrap" such nodes and
correctly detect links in bold and/or italic text.